+++ engine/client/cl_pred.cpp
@@ -40,7 +40,7 @@ void CL_CheckPredictionError (void)
 
 		// save for error itnerpolation
 		for (i=0 ; i<3 ; i++)
-			cl.prediction_error[i] = delta[i]*0.125;
+			cl.prediction_error[i] = delta[i]*0.125f;
 	}
 }
 
@@ -219,7 +219,7 @@ void CL_PredictMovement (void)
 	pm.trace = CL_PMTrace;
 	pm.pointcontents = CL_PMpointcontents;
 
-	pm_airaccelerate = atof(cl.configstrings[CS_AIRACCEL]);
+	pm_airaccelerate = (float)atof(cl.configstrings[CS_AIRACCEL]);
 
 	pm.s = cl.frame.playerstate.pmove;
 
@@ -245,15 +245,15 @@ void CL_PredictMovement (void)
 	step = pm.s.origin[2] - oldz;
 	if (step > 63 && step < 160 && (pm.s.pm_flags & PMF_ON_GROUND) )
 	{
-		cl.predicted_step = step * 0.125;
+		cl.predicted_step = step * 0.125f;
 		cl.predicted_step_time = cls.realtime - cls.frametime * 500;
 	}
 
 
 	// copy results out for rendering
-	cl.predicted_origin[0] = pm.s.origin[0]*0.125;
-	cl.predicted_origin[1] = pm.s.origin[1]*0.125;
-	cl.predicted_origin[2] = pm.s.origin[2]*0.125;
+	cl.predicted_origin[0] = pm.s.origin[0]*0.125f;
+	cl.predicted_origin[1] = pm.s.origin[1]*0.125f;
+	cl.predicted_origin[2] = pm.s.origin[2]*0.125f;
 
 	VectorCopy (pm.viewangles, cl.predicted_angles);
 }
